fix: react same key errors have been fixed by setting new keys using dynamic values.
the AddWordAttribute modal has a bug that closes the modal it's placed in when itself is closed by clicking on the inner modal. So, I've added a cancel button to close AddWordAttribute modal and am planning to remove this dismissing feature by clicking outside of the modal on this modal.